Fix state bug when reloading and update style
The state of 'topics.currentTopic' is set when clicking on a topic in
the Topics component, reloading the page while looking the threads of
a topic resets the state. Therefore, the fix checks if current topic
url is loaded onto the state and retrieves it from backend if it is
not.